Change Node-Layout to u32 again & Custom Option<Node>-Type

+++ src/repr/directed.rs
@@ -47,7 +47,7 @@ impl<OutNbs: Neighborhood> DirectedAdjacencyList for DirectedGraph<OutNbs> {
     fn in_neighbors_of(&self, u: Node) -> impl Iterator<Item = Node> + '_ {
         // Should be avoided as this is very costly
         self.vertices()
-            .filter(move |&v| self.out_nbs[v].has_neighbor(u))
+            .filter(move |&v| self.out_nbs[v as usize].has_neighbor(u))
     }
 
     fn in_degree_of(&self, u: Node) -> NumNodes {
@@ -58,17 +58,17 @@ impl<OutNbs: Neighborhood> DirectedAdjacencyList for DirectedGraph<OutNbs> {
 
 impl<OutNbs: Neighborhood> AdjacencyTest for DirectedGraph<OutNbs> {
     fn has_edge(&self, u: Node, v: Node) -> bool {
-        self.out_nbs[u].has_neighbor(v)
+        self.out_nbs[u as usize].has_neighbor(v)
     }
 
     fn has_neighbors<const N: usize>(&self, u: Node, neighbors: [Node; N]) -> [bool; N] {
-        self.out_nbs[u].has_neighbors(neighbors)
+        self.out_nbs[u as usize].has_neighbors(neighbors)
     }
 }
 
 impl<OutNbs: Neighborhood> GraphEdgeEditing for DirectedGraph<OutNbs> {
     fn try_add_edge(&mut self, u: Node, v: Node) -> bool {
-        if !self.out_nbs[u].try_add_neighbor(v) {
+        if !self.out_nbs[u as usize].try_add_neighbor(v) {
             self.num_edges += 1;
             false
         } else {
@@ -77,7 +77,7 @@ impl<OutNbs: Neighborhood> GraphEdgeEditing for DirectedGraph<OutNbs> {
     }
 
     fn try_remove_edge(&mut self, u: Node, v: Node) -> bool {
-        if self.out_nbs[u].try_remove_neighbor(v) {
+        if self.out_nbs[u as usize].try_remove_neighbor(v) {
             self.num_edges -= 1;
             true
         } else {
@@ -91,13 +91,13 @@ impl<OutNbs: Neighborhood> GraphDirectedEdgeEditing for DirectedGraph<OutNbs> {
         // Should be avoided as this is very costly
         self.num_edges -= self
             .vertices_range()
-            .map(|v| self.out_nbs[v].try_remove_neighbor(u) as NumEdges)
+            .map(|v| self.out_nbs[v as usize].try_remove_neighbor(u) as NumEdges)
             .sum::<NumEdges>();
     }
 
     fn remove_edges_out_of_node(&mut self, u: Node) {
-        self.num_edges -= self.out_nbs[u].num_of_neighbors() as NumEdges;
-        self.out_nbs[u].clear();
+        self.num_edges -= self.out_nbs[u as usize].num_of_neighbors() as NumEdges;
+        self.out_nbs[u as usize].clear();
     }
 }
 
@@ -112,15 +112,15 @@ impl<OutNbs: Neighborhood, InNbs: Neighborhood> DirectedAdjacencyList
     for DirectedGraphIn<OutNbs, InNbs>
 {
     fn in_neighbors_of(&self, u: Node) -> impl Iterator<Item = Node> + '_ {
-        self.in_nbs[u].neighbors()
+        self.in_nbs[u as usize].neighbors()
     }
 
     fn in_degree_of(&self, u: Node) -> NumNodes {
-        self.in_nbs[u].num_of_neighbors()
+        self.in_nbs[u as usize].num_of_neighbors()
     }
 
     fn in_neighbors_of_as_stream(&self, u: Node) -> impl BitmaskStream + '_ {
-        self.in_nbs[u].neighbors_as_stream(self.number_of_nodes())
+        self.in_nbs[u as usize].neighbors_as_stream(self.number_of_nodes())
     }
 }
 
@@ -129,14 +129,14 @@ impl<OutNbs: Neighborhood, InNbs: Neighborhood> AdjacencyTest for DirectedGraphI
         // Without additional knowledge, checking `out_nbs` or `in_nbs` does not make a difference:
         // since we define `AdjArrayMatrix` which uses a `BitNeighborhood` for `in_nbs`, we default
         // to `in_nbs` here
-        self.in_nbs[u].has_neighbor(v)
+        self.in_nbs[u as usize].has_neighbor(v)
     }
 
     fn has_neighbors<const N: usize>(&self, u: Node, neighbors: [Node; N]) -> [bool; N] {
         // Without additional knowledge, checking `out_nbs` or `in_nbs` does not make a difference:
         // since we define `AdjArrayMatrix` which uses a `BitNeighborhood` for `in_nbs`, we default
         // to `in_nbs` here
-        self.in_nbs[u].has_neighbors(neighbors)
+        self.in_nbs[u as usize].has_neighbors(neighbors)
     }
 }
 
@@ -144,7 +144,7 @@ impl<OutNbs: Neighborhood, InNbs: Neighborhood> GraphEdgeEditing
     for DirectedGraphIn<OutNbs, InNbs>
 {
     fn try_add_edge(&mut self, u: Node, v: Node) -> bool {
-        if !self.out_nbs[u].try_add_neighbor(v) {
+        if !self.out_nbs[u as usize].try_add_neighbor(v) {
             self.num_edges += 1;
             false
         } else {
@@ -153,7 +153,7 @@ impl<OutNbs: Neighborhood, InNbs: Neighborhood> GraphEdgeEditing
     }
 
     fn try_remove_edge(&mut self, u: Node, v: Node) -> bool {
-        if self.out_nbs[u].try_remove_neighbor(v) {
+        if self.out_nbs[u as usize].try_remove_neighbor(v) {
             self.num_edges -= 1;
             true
         } else {
@@ -168,19 +168,19 @@ impl<OutNbs: Neighborhood, InNbs: Neighborhood> GraphDirectedEdgeEditing
     fn remove_edges_into_node(&mut self, u: Node) {
         self.num_edges -= self
             .vertices_range()
-            .map(|v| self.out_nbs[v].try_remove_neighbor(u) as NumEdges)
+            .map(|v| self.out_nbs[v as usize].try_remove_neighbor(u) as NumEdges)
             .sum::<NumEdges>();
-        self.num_edges -= self.in_nbs[u].num_of_neighbors() as NumEdges;
-        self.in_nbs[u].clear();
+        self.num_edges -= self.in_nbs[u as usize].num_of_neighbors() as NumEdges;
+        self.in_nbs[u as usize].clear();
     }
 
     fn remove_edges_out_of_node(&mut self, u: Node) {
         self.num_edges -= self
             .vertices_range()
-            .map(|v| self.in_nbs[v].try_remove_neighbor(u) as NumEdges)
+            .map(|v| self.in_nbs[v as usize].try_remove_neighbor(u) as NumEdges)
             .sum::<NumEdges>();
-        self.num_edges -= self.out_nbs[u].num_of_neighbors() as NumEdges;
-        self.out_nbs[u].clear();
+        self.num_edges -= self.out_nbs[u as usize].num_of_neighbors() as NumEdges;
+        self.out_nbs[u as usize].clear();
     }
 }
